+++ src/templates/react-app/api/express/client/models/{{modelNameCamel}}.mustache.js
@@ -1,12 +1,7 @@
-import { get } from "../restClient";
-
-function sleep(ms) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
+import { post, get, put, remove } from "./helpers/restClient";
 
 export async function add{{modelNamePascal}}({{modelNameCamel}}) {
-  await sleep(1000);
-  throw Error('Not implemented');
+  return await post("{{modelNamePascal}}", {{modelNameCamel}});
 }
 
 export async function get{{modelNamePascal}}(id) {
@@ -14,16 +9,13 @@ export async function get{{modelNamePascal}}(id) {
 }
 
 export async function get{{pluralModelNamePascal}}() {
-  await sleep(1000);
-  throw Error('Not implemented');
+  return await get("{{pluralModelNamePascal}}");
 }
 
 export async function update{{modelNamePascal}}(id, {{modelNameCamel}}) {
-  await sleep(1000);
-  throw Error('Not implemented');
+  return await put("{{modelNamePascal}}", id, {{modelNameCamel}});
 }
 
 export async function remove{{modelNamePascal}}(id) {
-  await sleep(1000);
-  throw Error('Not implemented');
+  return await remove("{{modelNamePascal}}", id);
 }