+++ src/templates/react-app/react/base/src/components/common/MessageContainer.js
@@ -1,0 +1,15 @@
+import { Container } from "@material-ui/core";
+import { Header } from "./Header";
+
+export function MessageContainer({ title, message, className }) {
+  const classes = ["Container", className].join(" ");
+
+  return (
+    <Container maxWidth="sm" className={classes}>
+      <Header title={title} />
+      <div className="containerContent">
+        <h2>{message}</h2>
+      </div>
+    </Container>
+  );
+}